added lock dyes option to the skinning table

+++ src/main/java/riskyken/armourersWorkshop/utils/UtilItems.java
@@ -32,12 +32,12 @@ public final class UtilItems {
         stackNbt.setInteger(LibCommonTags.TAG_INTENSITY, intensity);
     }
     
-    public static void spawnItemInWorld(World world, int x, int y, int z, ItemStack stack) {
+    public static void spawnItemInWorld(World world, double x, double y, double z, ItemStack stack) {
         float f = 0.7F;
         double xV = (double)(world.rand.nextFloat() * f) + (double)(1.0F - f) * 0.5D;
         double yV = (double)(world.rand.nextFloat() * f) + (double)(1.0F - f) * 0.5D;
         double zV = (double)(world.rand.nextFloat() * f) + (double)(1.0F - f) * 0.5D;
-        EntityItem entityitem = new EntityItem(world, (double)x + xV, (double)y + yV, (double)z + zV, stack);
+        EntityItem entityitem = new EntityItem(world, x + xV, y + yV, z + zV, stack);
         world.spawnEntityInWorld(entityitem);
     }
 }
